+++ extension-manifest-v3/src/background/reporting/webrequest-monkey-patch.js
